今まではそれぞれのファイルにデータを作成していたので一つにするためにtweetDataというファイルを作成しました
