+++ server/src/infr/addTweet.ts
@@ -1,3 +1,5 @@
+import {tweetData} from "./tweetData";
+
 type TweetDataObj = {
   id: string;
   message: string;
@@ -6,31 +8,8 @@ type TweetDataObj = {
   userId: string;
 }
 
+const map: Record<string, TweetDataObj>  = tweetData()
 const date = new Date
-
-const map: Record<string, TweetDataObj> = {
-  1: {
-    id: "1",
-    message: "hello",
-    like: 0,
-    createdAt: Date.toLocaleString(),
-    userId: "tokitoki"
-  },
-  2: {
-    id: "2",
-    message: "world",
-    like: 0,
-    createdAt: Date.toLocaleString(),
-    userId: "tokitoki"
-  },
-  3: {
-    id: "3",
-    message: "japan",
-    like: 0,
-    createdAt: Date.toLocaleString(),
-    userId: "tokitoki"
-  }
-}
 
 export async function addTweet (userId: string, message: string): Promise<TweetDataObj>{
   const  id = String(Object.keys(map).length + 1)
