+++ server/src/infr/deleteTweet.ts
@@ -1,3 +1,5 @@
+import {tweetData} from "./tweetData"
+
 type TweetDataObj = {
   id: string;
   message: string;
@@ -6,31 +8,7 @@ type TweetDataObj = {
   userId: string;
 }
 
-const date = new Date
-
-const map: Record<string, TweetDataObj> = {
-  1: {
-    id: "1",
-    message: "hello",
-    like: 0,
-    createdAt: date.toLocaleString(),
-    userId: "tokitoki"
-  },
-  2: {
-    id: "2",
-    message: "world",
-    like: 0,
-    createdAt: date.toLocaleString(),
-    userId: "tokitoki"
-  },
-  3: {
-    id: "3",
-    message: "japan",
-    like: 0,
-    createdAt: date.toLocaleString(),
-    userId: "tokitoki"
-  }
-}
+const map: Record<string, TweetDataObj>  = tweetData()
 
 export async function deleteTweet(id: string): Promise<void | null> {
   if(map[id]) {
