+++ server/src/infr/getTweet.ts
@@ -1,3 +1,5 @@
+import {tweetData} from "./tweetData";
+
 type TweetDataObj = {
   id: string;
   message: string;
@@ -8,38 +10,7 @@ type TweetDataObj = {
 
 const date = new Date
 
-
-const map: Record<string, TweetDataObj> = {
-  1: {
-    id: "1",
-    message: "hello",
-    like: 0,
-    createdAt: date.toLocaleString(),
-    userId: "tokitoki"
-  },
-  2: {
-    id: "2",
-    message: "world",
-    like: 0,
-    createdAt: date.toLocaleString(),
-    userId: "tokitoki"
-  },
-  3: {
-    id: "3",
-    message: "japan",
-    like: 0,
-    createdAt: date.toLocaleString(),
-    userId: "tokitoki"
-  },
-  4: {
-    id: "4",
-    message: "ときです",
-    like: 0,
-    createdAt: date.toLocaleString(),
-    userId: "YoshinoToki"
-  },
-}
-
+const map: Record<string, TweetDataObj>  = tweetData()
 
 export async function getTweet(catchId: string): Promise<TweetDataObj | null> {
   if(map[catchId]) {
